requirements.txt
attrs==23.1.0
aws-cdk.assets==1.203.0
aws-cdk.aws-acmpca==1.203.0
aws-cdk.aws-applicationautoscaling==1.203.0
aws-cdk.aws-autoscaling-common==1.203.0
aws-cdk.aws-certificatemanager==1.203.0
aws-cdk.aws-cloudformation==1.203.0
aws-cdk.aws-cloudfront==1.203.0
aws-cdk.aws-cloudwatch==1.203.0
aws-cdk.aws-codeguruprofiler==1.203.0
aws-cdk.aws-codestarnotifications==1.203.0
aws-cdk.aws-ec2==1.203.0
aws-cdk.aws-ecr==1.203.0
aws-cdk.aws-ecr-assets==1.203.0
aws-cdk.aws-efs==1.203.0
aws-cdk.aws-events==1.203.0
aws-cdk.aws-iam==1.203.0
aws-cdk.aws-kms==1.203.0
aws-cdk.aws-lambda==1.203.0
aws-cdk.aws-logs==1.203.0
aws-cdk.aws-route53==1.203.0
aws-cdk.aws-s3==1.203.0
aws-cdk.aws-s3-assets==1.203.0
aws-cdk.aws-s3-deployment==1.203.0
aws-cdk.aws-s3-notifications==1.203.0
aws-cdk.aws-signer==1.203.0
aws-cdk.aws-sns==1.203.0
aws-cdk.aws-sqs==1.203.0
aws-cdk.aws-ssm==1.203.0
aws-cdk.cloud-assembly-schema==1.203.0
aws-cdk.core==1.203.0
aws-cdk.custom-resources==1.203.0
aws-cdk.cx-api==1.203.0
aws-cdk.lambda-layer-awscli==1.203.0
aws-cdk.region-info==1.203.0
boto3==1.26.151
botocore==1.29.151
cachetools==5.3.1
cattrs==22.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
constructs==3.4.322
gnupg==2.3.1
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
idna==3.4
importlib-resources==5.12.0
jmespath==1.0.1
jsii==1.83.0
protobuf==4.23.2
psutil==5.9.5
publication==0.0.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
typeguard==2.13.3
typing_extensions==4.6.3
urllib3==1.26.16
